Se agregan excepciones

+++ src/main/java/com/example/demo/service/OdontologoService.java
@@ -1,8 +1,12 @@
 package com.example.demo.service;
 
+import com.example.demo.exceptions.ResourceCouldNotBeAdded;
+import com.example.demo.exceptions.ResourceNotFoundException;
 import com.example.demo.persistence.entities.Odontologo;
+import com.example.demo.persistence.entities.Paciente;
 import com.example.demo.persistence.repository.IOdontologoRepository;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpStatus;
 import org.springframework.stereotype.Service;
 
 import java.util.List;
@@ -14,32 +18,52 @@ public class OdontologoService{
     IOdontologoRepository odontologoRepository;
 
     //Methods
-    public List<Odontologo> listAll(){
-        return odontologoRepository.findAll();
+    public List<Odontologo> listAll() throws ResourceNotFoundException {
+
+        List<Odontologo> odontologos = odontologoRepository.findAll();
+
+        if (odontologos.isEmpty()) {
+            throw new ResourceNotFoundException("No se encontraron odontologos");
+        }
+
+        return odontologos;
     }
 
     public Odontologo findById(Long id){
         return odontologoRepository.findById(id).orElse(null);
     }
 
-    public Odontologo addOdontologo(Odontologo odontologo){
+    public Odontologo addOdontologo(Odontologo odontologo) throws ResourceCouldNotBeAdded {
+
+        if (odontologo.getNombre().equals("") || odontologo.getApellido().equals("") || odontologo.getMatricula().equals("")){
+            throw new ResourceCouldNotBeAdded("No se pudo agregar el odontólogo a la base de datos, el formato no es el correcto");
+        }
+
         return odontologoRepository.save(odontologo);
     }
 
-    public void deleteOdontologo(Long id){
+    public void deleteOdontologo(Long id) throws ResourceNotFoundException {
+
+        if (findById(id) == null){
+            throw new ResourceNotFoundException("No existe un odontologo con ese Id");
+        }
         odontologoRepository.deleteById(id);
     }
 
-    public void modifyOdontologo(Long id, String nombre, String apellido, String matricula) {
-        Odontologo odontologo = odontologoRepository.findById(id).orElse(null);
-        if (odontologo != null) {
-            odontologo.setNombre(nombre);
-            odontologo.setApellido(apellido);
-            odontologo.setMatricula(matricula);
-            odontologoRepository.save(odontologo);
+    public void modifyOdontologo(Long id, String nombre, String apellido, String matricula) throws ResourceNotFoundException{
+
+        if (findById(id) == null){
+            throw new ResourceNotFoundException("No existe un odontologo con ese Id");
         }
+
+        Odontologo odontologo = odontologoRepository.findById(id).orElse(null);
+
+        odontologo.setNombre(nombre);
+        odontologo.setApellido(apellido);
+        odontologo.setMatricula(matricula);
+
+        odontologoRepository.save(odontologo);
+
     }
-
-
 
 }
